feat(core/form/inputs): make editor instance and changes available outside the PT-input
This will allow the initializer of the input to get access to the editor and editor change events.

It is othewise a bit tricky to get hold of it, as we are outside the editor context at
the point when we are creating the input and can't use the normal hooks (`usePortableTextEditor`).

The editor will call a 'ready' change when it's mounted, which could be used to
set a editor reference in the component that renders the PortableTextInput.

Example:
```ts
  const handleEditorChange = useCallback((change: EditorChange, editor: PortableTextEditor) => {
    if (change.type === 'ready') {
      editorRef.current = editor
    }
  }, [])
```